requirements.txt
numpy
matplotlib
scipy
pyqt5
pyqtgraph
pyopengl
scikit-sparse